update(GiniCaptureSDK, GiniBankSDK): remove ComponentAPI
PIA-3348
